+++ src/inmanta/ast/type.py
@@ -17,8 +17,9 @@
 """
 
 import numbers
+from typing import Callable
 from typing import List as PythonList
-from typing import Optional
+from typing import Optional, Sequence
 
 from inmanta.ast import DuplicateException, Locatable, Location, Named, Namespace, RuntimeException, TypeNotFoundException
 from inmanta.execute.util import AnyType, NoneValue
@@ -126,7 +127,25 @@ class NullableType(Type):
         self.basetype.normalize()
 
 
-class Number(Type):
+class Primitive(Type):
+    def __init__(self) -> None:
+        Type.__init__(self)
+        self.try_cast_functions: Sequence[Callable[[Optional[object]], object]] = []
+
+    def cast(self, value: Optional[object]) -> object:
+        exception: RuntimeException = RuntimeException(None, "Failed to cast '%s' to %s" % (value, self.type_string()))
+
+        for cast in self.try_cast_functions:
+            try:
+                return cast(value)
+            except ValueError:
+                continue
+            except TypeError:
+                raise exception
+        raise exception
+
+
+class Number(Primitive):
     """
         This class represents an integer or float in the configuration model. On
         these integers the following operations are supported:
@@ -135,7 +154,8 @@ class Number(Type):
     """
 
     def __init__(self) -> None:
-        Type.__init__(self)
+        Primitive.__init__(self)
+        self.try_cast_functions: Sequence[Callable[[Optional[object]], numbers.Number]] = [int, float]
 
     def validate(self, value: Optional[object]) -> bool:
         """
@@ -160,13 +180,34 @@ class Number(Type):
         return "number"
 
 
-class Bool(Type):
+class Integer(Number):
+    """
+        An innstance of this class represent the int type in the configuration model.
+    """
+
+    def __init__(self) -> None:
+        Number.__init__(self)
+        self.try_cast_functions: Sequence[Callable[[Optional[object]], object]] = [int]
+
+    def validate(self, value: Optional[object]) -> bool:
+        if not super().validate(value):
+            return False
+        if not isinstance(value, numbers.Integral):
+            raise RuntimeException(None, "Invalid value '%s', expected %s" % (value, self.type_string()))
+        return True
+
+    def type_string(self) -> str:
+        return "int"
+
+
+class Bool(Primitive):
     """
         This class represents a simple boolean that can hold true or false.
     """
 
     def __init__(self) -> None:
-        Type.__init__(self)
+        Primitive.__init__(self)
+        self.try_cast_functions: Sequence[Callable[[Optional[object]], object]] = [bool]
 
     def validate(self, value: Optional[object]) -> bool:
         """
@@ -189,13 +230,14 @@ class Bool(Type):
         return None
 
 
-class String(Type):
+class String(Primitive):
     """
         This class represents a string type in the configuration model.
     """
 
     def __init__(self) -> None:
-        Type.__init__(self)
+        Primitive.__init__(self)
+        self.try_cast_functions: Sequence[Callable[[Optional[object]], object]] = [str]
 
     def validate(self, value: Optional[object]) -> bool:
         """
@@ -208,6 +250,13 @@ class String(Type):
             raise RuntimeException(None, "Invalid value '%s', expected String" % value)
 
         return True
+
+    def cast(self, value: Optional[object]) -> object:
+        if value is True:
+            return "true"
+        if value is False:
+            return "false"
+        return super().cast(value)
 
     def type_string(self) -> str:
         return "string"
@@ -476,4 +525,11 @@ def create_function(expression: "ExpressionStatement"):
     return function
 
 
-TYPES = {"string": String(), "number": Number(), "bool": Bool(), "list": LiteralList(), "dict": LiteralDict()}
+TYPES = {
+    "string": String(),
+    "number": Number(),
+    "int": Integer(),
+    "bool": Bool(),
+    "list": LiteralList(),
+    "dict": LiteralDict(),
+}